+++ backend/src/main/java/com/wirewave/wirewave/repository/CommentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    List<Comment> findByProductId(Long productId);
-    List<Comment> findByUserId(Long userId);
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+    List<Comment> findByProductId(Integer productId);
+    List<Comment> findByUserId(Integer userId);
 }
